Add 16 and 8 bit version to simd implementation

+++ bitConvert.hpp
@@ -3,7 +3,7 @@
 
 void bitConvert_8(char * inputString, char **vector);
 void bitConvert_16(char * inputString, short int **vector);
-void bitConvert_32(char * inputString, int **vector);
+void bitConvert_32(char * inputString, long **vector);
 void bitConvert_64(char * inputString, long **vector);
 
 #endif
